requirements.txt
anthropic==0.40.0
pydantic==2.10.5
python-dotenv==1.0.1

# Testing
pytest==8.3.4
pytest-asyncio==0.25.2
